requirements.txt
slackclient==2.2.1
pymongo==3.9.0
mongoengine==0.18.2
python-dateutil==2.8.0
requests==2.22.0
uvloop==0.13.0

